refactor: improve module interfacing (#25)
This PR refactors 

1. the data meta-model module interface
2. the default Facade-X data meta-model according to this interface
3. the GeoSPARQL output to be according to this interface

(1) prepares for future modules by enabling more encapsulation of a
module class.

(2) improves the locality of the Facade-X output. 
That in turn makes it easier to make new modules based off the Facade-X
module.

(3) makes clearer how geometries are generated, when they're reprojected
and how they're output.